Move filter stuff to argus.filter
Keeps the model in place but does move the factory.
