+++ src/argus/filter/constants.py
