+++ src/argus/filter/primitive_serializers.py
