+++ src/argus/incident/V1/views.py
@@ -7,10 +7,10 @@ from rest_framework.generics import get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from ..filters import SourceLockedIncidentFilter
+from argus.filter.filters import SourceLockedIncidentFilter
+
 from ..models import Incident, SourceSystem
 from ..serializers import IncidentPureDeserializer, SourceSystemSerializer
-
 from ..views import IncidentViewSet, BooleanStringOAEnum
 from .serializers import (
     AcknowledgementSerializerV1,
