+++ src/argus/incident/views.py
@@ -34,12 +34,12 @@ from argus.notificationprofile.media import (
     send_notifications_to_users,
     background_send_notification,
 )
+from argus.filter.filters import IncidentFilter, SourceLockedIncidentFilter
 from argus.util.datetime_utils import INFINITY_REPR
 from argus.util.signals import bulk_changed
 from argus.util.utils import import_class_from_dotted_path
 
 from .forms import AddSourceSystemForm
-from .filters import IncidentFilter, SourceLockedIncidentFilter
 from .models import (
     ChangeEvent,
     Event,
