+++ src/argus/notificationprofile/checks.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.checks import Warning
 
-from .validators import validate_jsonfilter
+from argus.filter.validators import validate_jsonfilter
 
 
 __all__ = ["fallback_filter_check"]
