+++ src/argus/notificationprofile/models.py
@@ -13,8 +13,8 @@ from django.utils import timezone
 from django.utils.text import slugify
 
 from argus.auth.models import User
-
-from .constants import DEPRECATED_FILTER_NAMES
+from argus.filter.filterwrapper import FilterWrapper
+from argus.filter.constants import DEPRECATED_FILTER_NAMES
 
 if TYPE_CHECKING:
     from argus.incident.models import Event, Incident
@@ -100,77 +100,6 @@ class TimeRecurrence(models.Model):
         super().save(*args, **kwargs)
 
 
-class FilterWrapper:
-    TRINARY_FILTERS = ("open", "acked", "stateful")
-
-    def __init__(self, filterblob):
-        self.fallback_filter = getattr(settings, "ARGUS_FALLBACK_FILTER", {})
-        self.filter = filterblob.copy()
-
-    def _get_tristate(self, tristate):
-        fallback_filter = self.fallback_filter.get(tristate, None)
-        return self.filter.get(tristate, fallback_filter)
-
-    def are_tristates_empty(self):
-        for tristate in self.TRINARY_FILTERS:
-            if self._get_tristate(tristate) is not None:
-                return False
-        return True
-
-    def is_maxlevel_empty(self):
-        fallback_filter = self.fallback_filter.get("maxlevel", None)
-        return not self.filter.get("maxlevel", fallback_filter)
-
-    def is_event_types_empty(self):
-        fallback_filter = self.fallback_filter.get("event_types", None)
-        return not self.filter.get("event_types", fallback_filter)
-
-    def are_source_system_ids_empty(self):
-        fallback_filter = self.fallback_filter.get("sourceSystemIds", None)
-        return not self.filter.get("sourceSystemIds", fallback_filter)
-
-    def are_tags_empty(self):
-        fallback_filter = self.fallback_filter.get("tags", None)
-        return not self.filter.get("tags", fallback_filter)
-
-    @property
-    def is_empty(self):
-        return (
-            self.are_source_system_ids_empty()
-            and self.are_tags_empty()
-            and self.are_tristates_empty()
-            and self.is_maxlevel_empty()
-            and self.is_event_types_empty()
-        )
-
-    def get_incident_tristate_checks(self, incident) -> Dict[str, TriState]:
-        if self.are_tristates_empty():
-            return {}
-        fits_tristates = {}
-        for tristate in self.TRINARY_FILTERS:
-            filter_tristate = self._get_tristate(tristate)
-            if filter_tristate is None:
-                LOG.debug('Tristates: "%s" not in filter, ignoring', tristate)
-                fits_tristates[tristate] = None
-                continue
-            incident_tristate = getattr(incident, tristate, None)
-            LOG.debug('Tristates: "%s": filter = %s, incident = %s', tristate, filter_tristate, incident_tristate)
-            fits_tristates[tristate] = filter_tristate == incident_tristate
-        return fits_tristates
-
-    def incident_fits_maxlevel(self, incident):
-        if self.is_maxlevel_empty():
-            return None
-        fallback_filter = self.fallback_filter.get("maxlevel", None)
-        return incident.level <= min(filter(None, (self.filter["maxlevel"], fallback_filter)))
-
-    def event_fits(self, event):
-        if self.is_event_types_empty():
-            return True
-        fallback_filter = self.fallback_filter.get("event_types", None)
-        return event.type in self.filter.get("event_types", fallback_filter)
-
-
 class Filter(models.Model):
     FILTER_NAMES = set(DEPRECATED_FILTER_NAMES)
     user = models.ForeignKey(to=User, on_delete=models.CASCADE, related_name="filters")
@@ -198,6 +127,7 @@ class Filter(models.Model):
 
         return Incident.objects.all()
 
+    # XXX wrong location
     @property
     def filtered_incidents(self):
         if self.is_empty:
@@ -210,6 +140,7 @@ class Filter(models.Model):
 
         return filtered_by_source & filtered_by_tags & filtered_by_tristates & filtered_by_maxlevel
 
+    # XXX wrong location
     def incidents_with_source_systems(self, data=None):
         if not data:
             data = self.filter.copy()
@@ -218,6 +149,7 @@ class Filter(models.Model):
             return self.all_incidents.filter(source__in=source_list).distinct()
         return self.all_incidents.distinct()
 
+    # XXX wrong location
     def source_system_fits(self, incident: Incident, data=None):
         if not data:
             data = self.filter.copy()
@@ -227,6 +159,7 @@ class Filter(models.Model):
             return None
         return incident.source.id in source_list
 
+    # XXX wrong location
     def incidents_with_tags(self, data=None):
         if not data:
             data = self.filter.copy()
@@ -235,6 +168,7 @@ class Filter(models.Model):
             return self.all_incidents.from_tags(*tags_list)
         return self.all_incidents.distinct()
 
+    # XXX wrong location
     def tags_fit(self, incident: Incident, data=None):
         if not data:
             data = self.filter.copy()
@@ -245,6 +179,7 @@ class Filter(models.Model):
         tags = set(tag.representation for tag in incident.deprecated_tags)
         return tags.issuperset(tags_list)
 
+    # XXX wrong location
     def incidents_fitting_tristates(
         self,
         data=None,
@@ -270,6 +205,7 @@ class Filter(models.Model):
             fitting_incidents = fitting_incidents.stateless()
         return fitting_incidents.distinct()
 
+    # XXX wrong location
     def incidents_fitting_maxlevel(self, data=None):
         if not data:
             data = self.filter.copy()
@@ -278,6 +214,7 @@ class Filter(models.Model):
             return self.all_incidents.distinct()
         return self.all_incidents.filter(level__lte=maxlevel).distinct()
 
+    # XXX wrong location
     def incident_fits(self, incident: Incident):
         if self.is_empty:
             return False  # Filter is empty!
@@ -296,6 +233,7 @@ class Filter(models.Model):
             LOG.debug("Filter: %s: HIT!", self)
         return not any_failed
 
+    # XXX wrong location
     def event_fits(self, event: Event):
         if self.is_empty:
             return False  # Filter is empty!
@@ -374,11 +312,13 @@ class NotificationProfile(models.Model):
             return f"{self.name}"
         return f"{self.timeslot}: {', '.join(str(f) for f in self.filters.all())}"
 
+    # XXX wrong location
     @property
     def filtered_incidents(self):
         qs = [filter_.filtered_incidents for filter_ in self.filters.all()]
         return reduce(or_, qs)
 
+    # XXX wrong location
     def incident_fits(self, incident: Incident):
         assert incident.source, "incident does not have a source -2"
         if not self.active:
@@ -388,6 +328,7 @@ class NotificationProfile(models.Model):
         is_selected_by_filters = False not in checks.values()
         return is_selected_by_time and is_selected_by_filters
 
+    # XXX wrong location
     def event_fits(self, event: Event):
         if not self.active:
             return False
