+++ tests/notificationprofile/V1/test_views.py
@@ -6,9 +6,9 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 from rest_framework.test import APIClient
 
+from argus.filter.factories import FilterFactory
 from argus.notificationprofile.factories import (
     DestinationConfigFactory,
-    FilterFactory,
     NotificationProfileFactory,
     TimeslotFactory,
 )
